Grammar improvements, bug fixes and tests for regex
