--- project/gql/interpreter/types/finite_automaton.py
@@ -1,97 +1,0 @@
-from project.gql.interpreter.types.automaton import Automaton
-from project.gql.interpreter.types.set import Set
-from project.automaton_tools import get_nfa_from_graph, set_nfa_states, add_nfa_states
-from project.matrix_tools import BooleanAdjacencies
-
-from networkx import MultiDiGraph
-from pyformlang.finite_automaton import NondeterministicFiniteAutomaton
-
-from project.gql.interpreter.exceptions import NotImplementedException
-
-
-class FiniteAutomaton(Automaton):
-    def __init__(self, nfa: NondeterministicFiniteAutomaton):
-        self.nfa = nfa
-
-    def __str__(self):
-        return str(self.nfa.to_dict())
-
-    @classmethod
-    def from_graph(cls, graph: MultiDiGraph):
-        return cls(get_nfa_from_graph(graph))
-
-    def intersect(self, other: "FiniteAutomaton"):
-        left_nfa = BooleanAdjacencies(self.nfa)
-        right_nfa = BooleanAdjacencies(other.nfa)
-
-        intersection_nfa = left_nfa.intersect(right_nfa).to_nfa()
-
-        return FiniteAutomaton(intersection_nfa)
-
-    def union(self, other: "FiniteAutomaton"):
-        return FiniteAutomaton(self.nfa.union(other.nfa).to_deterministic())
-
-    def concatenate(self, other: "FiniteAutomaton"):
-        left_regex = self.nfa.to_regex()
-        right_regex = other.nfa.to_regex()
-
-        return FiniteAutomaton(
-            left_regex.concatenate(right_regex).to_epsilon_nfa().to_deterministic()
-        )
-
-    def inverse(self):
-        inverted_nfa = self.nfa.copy()
-
-        for state in inverted_nfa.states:
-            inverted_nfa.add_final_state(state)
-
-        for state in self.nfa.final_states:
-            inverted_nfa.remove_final_state(state)
-
-        return FiniteAutomaton(inverted_nfa)
-
-    def kleene(self):
-        return FiniteAutomaton(self.nfa.kleene_star().to_deterministic())
-
-    def set_start(self, start_states: Set):
-        self.nfa = set_nfa_states(self.nfa, start_states=start_states.set)
-
-    def set_final(self, final_states: Set):
-        self.nfa = set_nfa_states(self.nfa, final_states=final_states.set)
-
-    def add_start(self, start_states: Set):
-        self.nfa = add_nfa_states(self.nfa, start_states=start_states.set)
-
-    def add_final(self, final_states: Set):
-        self.nfa = add_nfa_states(self.nfa, final_states=final_states.set)
-
-    def get_reachable(self):
-        raise NotImplementedException("TODO")
-
-    @property
-    def start(self):
-        return Set(self.nfa.start_states)
-
-    @property
-    def final(self):
-        return Set(self.nfa.final_states)
-
-    @property
-    def labels(self):
-        return Set(self.nfa.symbols)
-
-    @property
-    def edges(self):
-        edges_set = set()
-
-        edges_dict = self.nfa.to_dict()
-        for u in edges_dict.keys():
-            for label, v_set in edges_dict.get(u).items():
-                for v in v_set:
-                    edges_set.add((u, label, v))
-
-        return Set(edges_set)
-
-    @property
-    def vertices(self):
-        return Set(self.nfa.states)
