+++ project/gql/interpreter/visitor.py
@@ -3,7 +3,7 @@ from project.gql.parser.antlr.gqlParser import gqlParser
 
 from project.gql.interpreter.types.type import Type
 from project.gql.interpreter.types.automaton import Automaton
-from project.gql.interpreter.types.regex import Regex
+from project.gql.interpreter.types.finite_automaton_regex import Regex
 from project.gql.interpreter.types.bool import Bool
 from project.gql.interpreter.types.set import Set
 
@@ -41,13 +41,19 @@ class Visitor(gqlVisitor):
             self.memory.add_variable(name, value)
 
     def visitExpr(self, ctx: gqlParser.ExprContext) -> Type:
+        unary_operations = {"NOT": "inverse", "KLEENE": "kleene"}
         binary_operations = {
             "AND": "intersect",
             "OR": "union",
             "DOT": "concatenate",
             "IN": "find",
         }
-        unary_operations = {"NOT": "inverse", "KLEENE": "kleene"}
+
+        for un_op in unary_operations:
+            if getattr(ctx, un_op)():
+                _type = self.visit(ctx.expr(0))
+
+                return getattr(_type, unary_operations[un_op])()
 
         for bin_op in binary_operations:
             if getattr(ctx, bin_op)():
@@ -58,12 +64,6 @@ class Visitor(gqlVisitor):
                     left_type, right_type = right_type, left_type
 
                 return getattr(left_type, binary_operations[bin_op])(right_type)
-
-        for un_op in unary_operations:
-            if getattr(ctx, un_op)():
-                _type = self.visit(ctx.expr(0))
-
-                return getattr(_type, unary_operations[un_op])()
 
         return self.visitChildren(ctx)
